requirements.txt
# pip install git+https://github.com/m-bain/whisperX.git
# pip install python-dotenv
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
alembic==1.15.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
asteroid-filterbanks==0.4.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
av==14.2.0
backcall==0.2.0
beautifulsoup4==4.13.4
bleach==6.2.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.9.0
contourpy==1.3.1
ctranslate2==4.5.0
cycler==0.12.1
decorator==5.2.1
defusedxml==0.7.1
diskcache==5.6.3
docopt==0.6.2
einops==0.8.1
exceptiongroup==1.3.0
executing==2.2.0
fastapi==0.115.12
faster-whisper==1.1.1
fastjsonschema==2.21.1
ffmpeg-python==0.2.0
filelock==3.13.1
flatbuffers==25.2.10
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.6.1
future==1.0.0
greenlet==3.1.1
h11==0.16.0
huggingface-hub==0.29.3
humanfriendly==10.0
HyperPyYAML==1.2.2
idna==3.10
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
julius==0.2.7
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kiwisolver==1.4.8
lightning==2.5.1
lightning-utilities==0.14.2
llama_cpp_python==0.3.8
Mako==1.3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
mpmath==1.3.0
multidict==6.2.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
networkx==3.3
nltk==3.9.1
numpy==2.2.5
omegaconf==2.3.0
onnxruntime==1.19.0
optuna==4.2.1
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pickleshare==0.7.5
pillow==11.0.0
pipreqs==0.5.0
platformdirs==4.3.8
primePy==1.3
prompt_toolkit==3.0.51
propcache==0.3.1
protobuf==6.30.2
pure_eval==0.2.3
pyannote.audio==3.3.2
pyannote.core==5.0.0
pyannote.database==5.1.3
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.1
pyparsing==3.2.3
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytorch-lightning==2.5.1
pytorch-metric-learning==2.8.1
pytz==2025.2
pywin32==310
PyYAML==6.0.2
pyzmq==26.4.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.24.0
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
semver==3.0.4
sentencepiece==0.2.0
shellingham==1.5.4
silero-vad==5.1.2
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.13.1
soupsieve==2.7
speechbrain==1.0.2
SQLAlchemy==2.0.40
stack-data==0.6.3
starlette==0.46.2
sympy==1.13.1
tabulate==0.9.0
tensorboardX==2.6.2.2
threadpoolctl==3.6.0
tinycss2==1.4.0
tokenizers==0.21.1
torch==2.5.1+cu121
torch-audiomentations==0.12.0
torch_pitch_shift==1.2.5
torchaudio==2.5.1+cu121
torchmetrics==1.7.0
torchvision==0.20.1+cu121
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.50.3
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.34.2
wcwidth==0.2.13
webencodings==0.5.1
whisperx @ git+https://github.com/m-bain/whisperX.git@f10dbf6ab1717e84db7733df9c0b21658ee68f9b
yarg==0.1.9
yarl==1.18.3